+++ include/list/list.h
@@ -13,7 +13,7 @@
 #include <stdbool.h>
 
 typedef void *list_t;
-typedef uint8_t list_element_t; // TODO: change uint8_t to generic type
+typedef int list_element_t; // TODO: change unsigned int to generic type
 typedef enum {DEFAULT_LIST, ARRAY_LIST, LINKED_LIST, DOUBLY_LINKED_LIST} list_type_t;
 
 /**
@@ -21,12 +21,12 @@ typedef enum {DEFAULT_LIST, ARRAY_LIST, LINKED_LIST, DOUBLY_LINKED_LIST} list_ty
  * possible to make an undefined length list.
  * @return status code indicating success of operation. If -1, the call failed.
  */
-int8_t list_init(list_t *list, uint8_t n, list_type_t type);
+int list_init(list_t *list, unsigned int n, list_type_t type);
 
 /**
  * Size
  */
-uint8_t list_length(const list_t list);
+unsigned int list_length(const list_t list);
 bool list_is_empty(const list_t list);
 bool list_is_full(const list_t list);
 
@@ -35,23 +35,23 @@ bool list_is_full(const list_t list);
  */
 list_element_t list_first(const list_t list);
 list_element_t list_last(const list_t list);
-list_element_t list_get(const list_t list, uint8_t n);
+list_element_t list_get(const list_t list, unsigned int n);
 
 /**
  * Insert
  */
-int8_t list_insert(list_t list, uint8_t n, list_element_t elem);
-int8_t list_append(list_t list, list_element_t elem);
-int8_t list_prepend(list_t list, list_element_t elem);
+int list_insert(list_t list, unsigned int n, list_element_t elem);
+int list_append(list_t list, list_element_t elem);
+int list_prepend(list_t list, list_element_t elem);
 
 /**
  * Remove
  */
-list_element_t list_remove(list_t list, uint8_t n);
+list_element_t list_remove(list_t list, unsigned int n);
 void list_clear(list_t list);
 
-int8_t list_locate(const list_t list, list_element_t elem,
-                   int8_t(*cmp)(list_element_t a, list_element_t b));
+int list_locate(const list_t list, list_element_t elem,
+                   int(*cmp)(list_element_t a, list_element_t b));
 
 /**
  * Destroy the list `list`. It must be called when the list won't be
