+++ include/stack/dynamic_stack.h
@@ -5,19 +5,19 @@
 #include <stdbool.h>
 
 typedef void* dynamic_stack_t;
-typedef uint8_t dynamic_stack_element_t;
+typedef unsigned int dynamic_stack_element_t;
 
 /**
  * Initialize a dynamic stack
  * 
  * @return positive integer if success, -1 if operation fails.
  */
-int8_t dynamic_stack_init(dynamic_stack_t *dstack);
+int dynamic_stack_init(dynamic_stack_t *dstack);
 
 /**
  * Size
  */
-uint8_t dynamic_stack_length(const dynamic_stack_t dstack);
+unsigned int dynamic_stack_length(const dynamic_stack_t dstack);
 bool dynamic_stack_is_empty(const dynamic_stack_t dstack);
 bool dynamic_stack_is_full(const dynamic_stack_t dstack);
 
@@ -25,7 +25,7 @@ bool dynamic_stack_is_full(const dynamic_stack_t dstack);
  * Stack operations
  */
 dynamic_stack_element_t dynamic_stack_top(const dynamic_stack_t dstack);
-int8_t dynamic_stack_push(dynamic_stack_t dstack, dynamic_stack_element_t elem);
+int dynamic_stack_push(dynamic_stack_t dstack, dynamic_stack_element_t elem);
 dynamic_stack_element_t dynamic_stack_pop(dynamic_stack_t dstack);
 
 void dynamic_stack_clear(dynamic_stack_t dstack);
