+++ include/stack/static_stack.h
@@ -5,19 +5,19 @@
 #include <stdbool.h>
 
 typedef void* static_stack_t;
-typedef uint8_t static_stack_element_t;
+typedef unsigned int static_stack_element_t;
 
 /**
  * Initialize a static stack of `n` elements
  * 
  * @return positive integer if success, -1 if operation fails.
  */
-int8_t static_stack_init(static_stack_t *sstack, uint8_t n);
+int static_stack_init(static_stack_t *sstack, unsigned int n);
 
 /**
  * Size
  */
-uint8_t static_stack_length(const static_stack_t sstack);
+unsigned int static_stack_length(const static_stack_t sstack);
 bool static_stack_is_empty(const static_stack_t sstack);
 bool static_stack_is_full(const static_stack_t sstack);
 
@@ -25,7 +25,7 @@ bool static_stack_is_full(const static_stack_t sstack);
  * Stack operations
  */
 static_stack_element_t static_stack_top(const static_stack_t sstack);
-int8_t static_stack_push(static_stack_t sstack, static_stack_element_t elem);
+int static_stack_push(static_stack_t sstack, static_stack_element_t elem);
 static_stack_element_t static_stack_pop(static_stack_t sstack);
 
 void static_stack_clear(static_stack_t sstack);
